Only cache GET and HEAD requests
As a rule, we don't ever want to cache POST, PUT or DELETE responses.
This bug was highlighted when running performance tests with Gatling [1]
which has an HTTP cache switched on by default.

I was under the impression that POST, PUT and DELETE were uncachable.
According to the HTTP spec they can be.

[1] https://gatling.io