Add framework status and framework fields
Add two fields to the Framework model:
- framework: an enum identifying the framework this is associated with
             eg. for 'G-Cloud 7' framework would be 'gcloud'
- status: an enum identifying the current status of the framework
          iteration.

This change does not make use of these fields yet. The status field will
replace the expired flag but the code change will go through in a
separate pull request.

This uses native Enums for the framework and status fields. They are
well supported in SQLAlchemy now and are cleaner than a manual
constraint based solution.

One of the tests has started failing because of switched order. On
investigation the order was not deterministic. I've added serviceName to
the dummy services to make ordering deterministic.